requirements.txt
anyio==3.3.0
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
click==7.1.2
colorama==0.4.4
fastapi==0.65.2
fastapi-utils==0.2.1
greenlet==1.1.0
h11==0.12.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
idna==3.2
iniconfig==1.1.1
loguru==0.5.3
packaging==21.0
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2021.3
PyYAML==5.4.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.14
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
watchgod==0.7
websockets==9.1
win32-setctime==1.0.3
pyjwt
